chore(ci): Run GitHub Actions on `ubuntu-22.04-arm`
Builds on ARM might be faster (see [Linux arm64 hosted runners now available for free in public repositories (Public Preview) 🎉 · community · Discussion #148648](https://github.com/orgs/community/discussions/148648)),
this is a test to check in a real scenario. If it's faster, we'll keep it that way. If it's not, we'll revert this commit.