feat: Add `HEAD /v1/server/config`
Fixes #174.

Answering `412` without a precondition in the request would have been incorrect,
therefore the route requires `If-None-Match: *` to work properly. Otherwise, it
returns `428`.

When we will support caching, we will remove the `If-None-Match: *` requirement,
but for now it's the correct way to do it.

Since it's a `HEAD` request there is no response body with error details, which
is why using status codes semantically was a requirement here.