feat!: Store the dashboard address as a URL
Using a IP address or hostname like for the server address prevented one from defining a port
(useful for local development) and prevented all sorts of possible scenarios.

Using a URL simplifies a lot of things.

This commit also updates the `demo` scenario to make it ready to send invitations
(the Dashboard URL was in a configuration key that was removed in #188).