fix(cron): Fix infinite loop when server config isn’t initialized
When "Not refreshing service accounts tokens: Server config not initialized." happens,
`continue` restarts the loop without a sleep. This means the check is made again,
but the server config is still not initialized, and the loop goes on forever flooding
the logs (and the CPU).

Service accounts tokens are generated at startup anyway so we can just start with a sleep.