Revert "Create release after tests pass (#218)"
This reverts commit 68241e2465ccfc8d8e964db7a9cd3359398a83c3.

Releases were broken, as GitHub Actions wouldn’t run on the `master` branch.
See https://github.com/prose-im/prose-pod-api/actions/runs/15034985796.
Let’s just revert to when it worked, and rely on the fact that the CI already runs tests.
If we break things by force-pushing to `master`, well, we asked for problems.