chore: Remove `Error::logged`
This logic was added at a time I was debugging Cucumber logging errors multiple times,
but the bug is internal to cucumber-rs and has nothing to do with our code.

`Error::log` is only called in `IntoResponse::into_response` which takes ownership
so we’re still sure it will only be called once.